update fence api and modify tool_bar

+++ app/src/main/java/project/boostcamp/final_project/Model/Constant.java
@@ -15,6 +15,6 @@ public final class Constant {
 
     public static final String PACKAGE_NAME = "project.boostcamp.final_project";
 
-    public static final float GEOFENCE_RADIUS_IN_METERS = 1609; // 1 mile, 1.6 km
+    public static final int GEOFENCE_RADIUS_IN_METERS = 1609; // 1 mile, 1.6 km
 
 }
